Disallow comments and annotate attribute items
Instead of allowing comments and using them to describe the filesystem
layout, use a better annotated entry format in the first place.

All directory entry attributes except for start and size are now
written as key=value pairs.

Start and size are written as @start+size which will hopefully be easy
to comprehend by the reader.

The key for the filename attribute is "P" in order to give a hint to
the user that this is a pathname and not just a basename.
Include file mode and mtime at FS generation time

Also support comments, except in the first two lines which have a
predetermined format.