Simplify on-disk format
No more per-file attributes or timestamps. There is only the creation
time stamp for the whole filesystem, which is shared by all of its
files.

File permissions are also pointless since there is no instance which
enforces them.

Special file types like block device, fifo and so on do not make any
sense either, because a trivial filesystem is not intended to be ever
mounted. Also, special files are not "trivial", defeating TFS's purpose.