Make raw operation the default
Instead of specifying option -r for raw operation, option -d needs to be
specified for dynamic size suballocation.

This decision has been made because dynamic suballocation is not
something which can be figured out just by looking at the TrivialFS
metadata. Instead, it is a per-file application-level choice.

Also, dynamic suballocation is not required by read-only files which
have been allocated with the correct sizes.

It is only useful by files which will be updated in-place after initial
creation.