Add alignment support to trivialfs_alloc
A new command "%<alignment>=<offset>" enables alignment verification of
manually specified offsets and lengths. It defines that all lengths must
be multiples of <alignment>, and that all offsts must be equivalent to
<offset> modulo <alignment>.

The commands for searching for a suitable gap in ascending or descending
direction are modified by the alignment specification as follows: First,
the specified length is silently rounded up to the next multiple of
<alignment>. In addition, only offsets compliant with the alignment
definition are considered as possible target positions.